backend/app/schemas/auth.py
from pydantic import BaseModel

from app.models.user import UserRole


class LoginRequest(BaseModel):
    email: str
    password: str


class TokenResponse(BaseModel):
    access_token: str
    token_type: str = "bearer"


class UserResponse(BaseModel):
    id: int
    name: str
    email: str
    role: UserRole
    team_id: int | None = None
    manager_id: int | None = None

    model_config = {"from_attributes": True}
