backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pyjwt>=2.8.0
bcrypt>=4.1.2
weasyprint>=62.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
